requirements.txt
PyPDF2==3.0.1
Pillow==10.0.1
pdf2image==1.16.3
reportlab==4.0.4
PyMuPDF==1.23.14
cx-Freeze==6.15.16
pyinstaller==6.3.0
